+++ src/CMQMaster.h
@@ -14,7 +14,9 @@ public:
     std::string listen(Rcpp::CharacterVector addrs) {
         sock = zmq::socket_t(*ctx, ZMQ_ROUTER);
         sock.set(zmq::sockopt::router_mandatory, 1);
+        #ifdef ZMQ_BUILD_DRAFT_API
         sock.set(zmq::sockopt::router_notify, ZMQ_NOTIFY_DISCONNECT);
+        #endif
 
         int i;
         for (i=0; i<addrs.length(); i++) {